pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["k1cker666 <pomoymu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.tgbot.dependencies]
psycopg = {extras = ["binary"], version = "^3.1.19"}
psycopg-pool = "^3.2.2"
redis = {extras = ["hiredis"], version = "^5.0.4"}
pydantic = "^2.7.1"
loguru = "^0.7.2"
pillow = "^10.3.0"
pytest = "^8.2.0"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.migrations.dependencies]
psycopg = {extras = ["binary"], version = "^3.1.19"}
loguru = "^0.7.2"
click = "^8.1.7"
python-dotenv = "^1.0.1"


[tool.poetry.group.backend.dependencies]
psycopg = {extras = ["binary"], version = "^3.1.19"}
loguru = "^0.7.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
python-telegram-bot = "^21.1.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
redis = {extras = ["hiredis"], version = "^5.0.4"}
psycopg-pool = "^3.2.2"
pydantic = "^2.7.1"
loguru = "^0.7.2"
pillow = "^10.3.0"
pytest = "^8.2.0"
click = "^8.1.7"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
